convert retry_aead_ctx to use aead_ctx_alloc/free
Using statically defined structs from external libraries that may be
built as DSO's is dangerous, as the run time linked library may define a
struct that is larger or smaller than the version of the struct that was
defined when the library was built.  This is why openssl only exports
public symbols as opaque structs which must be heap allocated.  Convert
the retry_aead_ctx variable to use the new per-platform heap allocation
methods