remove needless test in bin/prog.c
prog_new_session_cb makes a call to SSL_CTX_set_tlsext_ticket_keys,
which passes on boringssl and fails on openssl due to the fact that the
ticket length that is required between the two differs.

The call is documented as a test only (and is clearly so because it
passes in an array of zeros, ostensibly just to test that the call
works.  As such it seems a bit silly to preform, just remove it